+++ packages/vue-start/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tanstack/vue-start
 
+## 1.168.47
+
+### Patch Changes
+
+- Updated dependencies [[`2f20c00`](https://github.com/TanStack/router/commit/2f20c00224c5ba63467551914e0c37012588c4c2), [`28a5e45`](https://github.com/TanStack/router/commit/28a5e4504e4ea5cb1480667a4bea2588a53e110f), [`f0b5eda`](https://github.com/TanStack/router/commit/f0b5eda544606686a8a8d675a686ca1366428b96), [`37877da`](https://github.com/TanStack/router/commit/37877da166fe4ce055c7b85e138b6681ebd7e8b4), [`0497cae`](https://github.com/TanStack/router/commit/0497caeef3ff7e1c1c6080eca38bca24e7ec320b), [`0caf6b9`](https://github.com/TanStack/router/commit/0caf6b9a2b7e14b0b146c74cc27cb05c19d700a5)]:
+  - @tanstack/vue-router@1.170.30
+  - @tanstack/start-plugin-core@1.171.40
+  - @tanstack/start-client-core@1.170.28
+  - @tanstack/start-server-core@1.169.32
+  - @tanstack/vue-start-client@1.167.33
+  - @tanstack/vue-start-server@1.167.37
+
 ## 1.168.46
 
 ### Patch Changes